refactor: move font manager and view manager to view package
