+++ src/main/java/view/FontManager.java
@@ -1,4 +1,4 @@
-package utility;
+package view;
 
 import javax.swing.*;
 import java.awt.*;
