+++ src/main/java/view/ViewManager.java
@@ -1,6 +1,5 @@
-package utility;
+package view;
 
-import view.IComponent;
 import view.view_events.SwitchPanelEvent;
 import view.view_events.ViewEvent;
