+++ src/main/java/view/components/TableComponent.java
@@ -1,6 +1,6 @@
 package view.components;
 
-import utility.FontManager;
+import view.FontManager;
 
 import javax.swing.*;
 import javax.swing.table.DefaultTableModel;
