+++ src/main/java/view/panels/AssetPanel.java
@@ -1,8 +1,8 @@
 package view.panels;
 
-import utility.FontManager;
-import utility.ViewManager;
+import view.FontManager;
 import view.IComponent;
+import view.ViewManager;
 import view.view_events.UpdateAssetEvent;
 import view.view_events.ViewEvent;
