+++ src/main/java/view/panels/TradeSimulationPanel.java
@@ -1,8 +1,8 @@
 package view.panels;
 
-import utility.FontManager;
-import utility.ViewManager;
+import view.FontManager;
 import view.IComponent;
+import view.ViewManager;
 import view.components.ButtonComponent;
 import view.view_events.SwitchPanelEvent;
 import view.view_events.ViewEvent;
