+++ src/main/java/view/panels/TransactionHistoryPanel.java
@@ -2,9 +2,9 @@ package view.panels;
 
 import entity.Transaction;
 import entity.TransactionHistory;
-import utility.FontManager;
-import utility.ViewManager;
+import view.FontManager;
 import view.IComponent;
+import view.ViewManager;
 import view.components.ButtonComponent;
 import view.components.TableComponent;
 import view.view_events.SwitchPanelEvent;
